[TWEAK] Move files to lib
