+++ actionmailbox/app/controllers/action_mailbox/ingresses/amazon_ses/base_controller.rb
@@ -16,7 +16,9 @@ module ActionMailbox
 
         private
           def set_notification
-            @notification = SnsNotification.new(request.raw_post)
+            require "action_mailbox/amazon_ses/sns_notification"
+
+            @notification = ::ActionMailbox::AmazonSes::SnsNotification.new(request.raw_post)
           end
 
           def ensure_valid_topic
